+++ SDEngine/Source/SSAOManager.cpp
@@ -30,7 +30,7 @@ void SSAOManager::Init(int screenWidth, int screenHeight)
 {
 	ssaoRT = shared_ptr<RenderTexture>(new RenderTexture(screenWidth, screenHeight, TextureFormat::R32));
 
-	ssaoDownSampleRT = shared_ptr<RenderTexture>(new RenderTexture(screenWidth / 2.0f, screenHeight / 2.0f, TextureFormat::R32));
+	ssaoDownSampleRT = shared_ptr<RenderTexture>(new RenderTexture((float)screenWidth / 2.0f, (float)screenHeight / 2.0f, TextureFormat::R32));
 
 	ssaoDownSampleBlurRT = shared_ptr<RenderTexture>(new RenderTexture(screenWidth / 2.0f, screenHeight / 2.0f, TextureFormat::R32));
